thread_local for multi-threaded performance
